Created staging branch and finished desktop version of login page
This commit adds a staging branch (which serves as the primary
branch to commit changes to) while also creating  the layout
of the login screen on desktop devices.

Adding in some sort of icons to keep the page interesting
might be wise, but overall this should be a satisfactory starting
point.