Add PrivateRoute util
This commit contains some minor changes other than one
important one, with the important change being the creation
of a PrivateRoute component that can be used to conditionally
display a route based on whether the user is authenticated or
not.

Currently it is hardcoded to be true, but later this can be
set to validate via a function or redux/context.