Refactor to NextJS
This commit contains a large refactor from both the server
and next-client to support next.js as a client-side delivery.

The original client folder has been left in place for now
while further work is done to ensure the next client is fully
functional before deleting the old client from the repo.